+++ sdk/metricsadvisor/Azure.AI.MetricsAdvisor/src/GetAllFeedbackFilter.cs
@@ -13,19 +13,19 @@ namespace Azure.AI.MetricsAdvisor
     /// <see cref="MetricsAdvisorClient.GetAllFeedbackAsync"/> operations.
     /// </summary>
     [CodeGenModel("FeedbackDimensionFilter")]
-    public partial class GetAllFeedbackFilter
+    public partial class FeedbackFilter
     {
         /// <summary>
-        /// Initializes a new instance of the <see cref="GetAllFeedbackFilter"/> class.
+        /// Initializes a new instance of the <see cref="FeedbackFilter"/> class.
         /// </summary>
-        public GetAllFeedbackFilter()
+        public FeedbackFilter()
         {
         }
 
         /// <summary> Initializes a new instance of FeedbackDimensionFilter. </summary>
         /// <param name="dimension"> metric dimension filter. </param>
         /// <exception cref="ArgumentNullException"> <paramref name="dimension"/> is null. </exception>
-        internal GetAllFeedbackFilter(IDictionary<string, string> dimension)
+        internal FeedbackFilter(IDictionary<string, string> dimension)
         {
             if (dimension == null)
             {