+++ sdk/metricsadvisor/Azure.AI.MetricsAdvisor/src/Models/MetricFeedback/MetricPeriodFeedback.cs
@@ -11,7 +11,7 @@ namespace Azure.AI.MetricsAdvisor
     /// Feedback indicating that this is an interval of seasonality.
     /// </summary>
     [CodeGenModel("PeriodFeedback")]
-    [CodeGenSuppress(nameof(MetricPeriodFeedback), typeof(string), typeof(GetAllFeedbackFilter))]
+    [CodeGenSuppress(nameof(MetricPeriodFeedback), typeof(string), typeof(FeedbackFilter))]
     public partial class MetricPeriodFeedback : MetricFeedback
     {
         /// <summary>
@@ -40,7 +40,7 @@ namespace Azure.AI.MetricsAdvisor
         /// <param name="dimensionFilter"> . </param>
         /// <param name="valueInternal"> . </param>
         /// <exception cref="ArgumentNullException"> <paramref name="metricId"/>, <paramref name="dimensionFilter"/>, or <paramref name="valueInternal"/> is null. </exception>
-        internal MetricPeriodFeedback(string metricId, GetAllFeedbackFilter dimensionFilter, PeriodFeedbackValue valueInternal)
+        internal MetricPeriodFeedback(string metricId, FeedbackFilter dimensionFilter, PeriodFeedbackValue valueInternal)
             : base(metricId, dimensionFilter.DimensionKey)
         {
             if (valueInternal == null)